==> code/main/07_2_reconstruct.R <==
################################################################################
# Description: Reconstruct case counts from model-predicted deaths
# 
# Author: Emily S Nightingale
# Date created: 18/03/2021
# 
################################################################################
################################################################################

outdir <- "output/reconstruct"
figdir <- "figures/reconstruct"

# Specify quantiles for CFR scaling and plotting
scale_quants <- c(0.25, 0.75)
plot_quants <- c(0.01,0.25,0.75,0.99)

# Load posterior predictions from death model, given average values of 
# case-fatality risk factors
# sims_long <- as.data.frame(readRDS(here::here("output","predict","sims_long_avgcov.rds")))
# nsims <- dplyr::n_distinct(sims_long$variable)

# Load observed confirmed cases
cases <- readRDS(here::here("data","aggregated","cases.rds"))[[1]]
deaths <- readRDS(here::here("data","aggregated","deaths.rds"))[[1]]

# Sample of LAs to check reconstruction
# la_samp <- sample(unique(cases$lad19nm),9)
la_samp <- c("Allerdale","Cornwall","Liverpool","Newcastle-under-Lyme","North Somerset","Rother","Southwark","Three Rivers","Watford")
la_samp <- c("Cannock Chase","East Suffolk","Kirklees","Merton","Rotherham","South Holland","Stockport","Wigan","Wyre")

# ---------------------------------------------------------------------------- #
## Reconstruct cases from predicted deaths ##

# Four steps performed by helper functions in /utils:
# + [get_cfr] Calculate post-P2 CFR between cases and predicted deaths per week/LTLA given 
# assumed lag
# + [rescale_sims] Scale each posterior sample of death time series by estimates post-pillar 2 
# CFR (according to specified scale_quants)
# + [agg_sims] Aggregate posteriors over country/geography/LTLA
# + [plot_reconstr] Plot quantiles alongside observed cases

for (lag in c(7,14,21)){
  result <- reconstruct(cases, lag = lag, plot = T, suffix = "")
  saveRDS(result, here::here(outdir,paste0("reconstruct_lag",lag,".rds")))
  rm(result) 
}

# overall CFR quantiles for 1 week lag:
# period 1  4.11  2.62  6.75
# period 2  5.37  3.07  9.09


# ---------------------------------------------------------------------------- #
## Plot for lag 7 ##

reconstruct <- readRDS(here::here(outdir, "reconstruct_lag7.rds"))
plot_reconst(reconstruct7$la, 7, sample = la_samp, save = T, figdir = figdir, h = 1200, w = 1800, res = 200)
plot_reconst(reconstruct$total, 7, save = T, figdir = figdir, title = F, caption = F) 
plot_reconst(reconstruct$geog, 7, save = T, figdir = figdir, title = F, caption = F)

# Combine for figure 3
plot_reconst(reconstruct$total, 7, save = F, figdir = figdir, title = F, caption = F) -> tot
plot_reconst(reconstruct$geog, 7, save = F, figdir = figdir, title = F, caption = F) -> geog
geog <- geog + labs(y = "") + guides(shape = "none")

png(here::here(figdir,"fig3.png"), height = 1200, width = 3200, res = 300)
# tiff(here::here("figures","paper","fig3.tif"), height = 1200, width = 3200, res = 300)
tot + geog + plot_annotation(tag_levels = "A")
dev.off()

###############################################################################
###############################################################################
